Majorly rework resource handling
This commit consists of multiple changes which couldn't be split
meaningfully without intermediary hacks:

* Replace chrisdew/protobuf with dcodeIO/ProtoBuf.js

The former doesn't work on io.js and seems inactive.

* Replace pre-generated files with a runtime generator

The steam_language*.js files would have to be re-generated anyway due to
a different protobuf API, but there is no way to represent "complex enum
properties", introduced in SteamKit2 after the last re-generation, as JS
object properties. Instead, use a JS port of SteamLanguageParser and
generate the Steam namespace on startup.

* Remove ref

It was only used for reading and writing 64-bit integers and C-strings
to/from Buffers. Instead, use ByteBuffer.js, since it's a dependency of
ProtoBuf.js anyway.

* Rewrite steamID.js

Use Long.js instead of ref and implement an API similar to ProtoBuf.js.

Consequences of the above changes:

- Steam resources (Protobufs and SteamLanguage) are now pulled from the
  SteamKit2 repo using a prepublish script. It uses svn because cloning
  the whole repo takes too long.

- _send now expects either a ByteBuffer or a ProtoBuf.js Message, to
  avoid having to call toBuffer in every method. Conversely, handlers
  now receive ByteBuffers for convenience.