rework sending and receiving messages
1. Applying protoMask to mark protobuf messages is ugly. And _send needs
to know whether the message is protobuf-backed, so why can't it apply
the mask itself? Also, abstracting the header from handlers works in all
cases except unified messages, which use the eresult and target_job_name
fields. To solve both problems, pass the parsed message header in both
directions (for now, just sending) instead of EMsg. If the "proto"
property (which is the proto header) is present, consider the message
protobuf-backed.

2. The current JobID system is ugly. To make a request, you have to call
_send with a function, even though it doesn't actually care that it's a
function - it just passes it back as-is with the response. Instead, use
the following symmetrical API: for incoming requests, provide a callback
that should be called with the response instead of _send; for outgoing
requests, a callback should be passed to _send, which will be called
directly with the response instead of looking for a handler. As a bonus,
it will also support the (currently unused?) case when a message is both
a request and a response.