Move most API into handler classes
Rationale:

1. Instead of a single huge README, each handler class can have its own.
It could be split anyway, but it's more intuitive when one class is
documented in a single file.

2. Extensibility. Anyone wishing to use a message that isn't implemented
has had to submit a feature request or resort to using the undocumented
_send method. Simply exposing it would create a problem for extension
modules: adding methods to SteamClient's prototype might lead to name
clashes, but exporting wrapper classes would be inconsistent with the
core methods.

3. A global message map doesn't work with unified messages. Every
handler function receives the same arguments, but only unified messages
need the header. Handler classes can listen for 'message' and pass on
whatever is needed to their own message maps.

To decouple user.js and steam_client.js, connection was split from
logon. Due to resulting implementation and documentation issues and
limited usefulness, it no longer automatically reconnects after
encryption handshake.

In addition, sending and receiving client messages was reworked too:

1. A header object instead of EMsg. This exposes the protobuf header for
messages that need it and provides a more convenient way to mark
protobuf messages than " | protoMask".

2. A symmetrical callback-based JobID wrapper. It supports (theoretical)
messages that are both a request and a response, and responses to client
requests are now handled in-place rather than through a separate
handler.